feature(Refactor) – Move interface and type declarations from /lib to src root directory
