+++ src/components/SliderItem.tsx
@@ -1,4 +1,4 @@
-import { Social, Technology, Tool } from '@/lib/interfaces';
+import { Social, Technology, Tool } from '@/interfaces';
 import Link from 'next/link';
 import Image from 'next/image';
 import { ArrowTopRightOnSquareIcon } from '@heroicons/react/24/solid';
