+++ src/components/social/Social.tsx
@@ -1,4 +1,4 @@
-import { Social } from '@/lib/interfaces';
+import { Social } from '@/interfaces';
 import SliderItem from '@/components/SliderItem';
 
 export default function Social({ social }: { social: Social }) {
