+++ src/components/technology/Technology.tsx
@@ -1,4 +1,4 @@
-import { Technology } from '@/lib/interfaces';
+import { Technology } from '@/interfaces';
 import SliderItem from '@/components/SliderItem';
 
 export default function Technology({ technology }: { technology: Technology }) {
