+++ src/components/tool/Tool.tsx
@@ -1,4 +1,4 @@
-import { Tool } from '@/lib/interfaces';
+import { Tool } from '@/interfaces';
 import SliderItem from '@/components/SliderItem';
 
 export default function Tool({ tool }: { tool: Tool }) {
