+++ src/interfaces.ts
