+++ src/types.ts
@@ -1,1 +1,3 @@
 export type ControlBarAction = 'Avatar' | 'Back' | 'Up';
+
+export type Brand = 'Figma' | 'WebStorm' | 'GitHub';
